Now using the new analysis with narrower prior.
added corrplot code for data availabilty

Added all the code for testing whether the proportion of opposite trends changes if one only looks at the 'significant' trends. This seems to be the case. But what does an opposite trend really mean? Maybe this should be done at the level of taxon pairs instead